Sync redis helm charts and fix max memory issue (#38)
* Sync icm-redis-operator/charts/icm-redis-cluster with icm-helm-charts/redis-cluster

* Fix operator log level

* Add redisnode comments

* Renamed redis container from .Chart.Name to redis-node
Fix for max-memory command line parsing

* Bump alpine version

* Remove max-memory and max-memory-policy

* Remove ip option

* Fix redis operator chart

* Fix matchLabels

* Add extraArgs field

* Add extraArgs field

* Fixed metrics label matcher
Removed debug arg

* Update charts/icm-redis-cluster/values.yaml


* Removed unnecessary label (being added below)
Fixed max-memory/maxmemory mixup

+++ pkg/config/redis.go
@@ -40,7 +40,7 @@ type Redis struct {
 	ServerBin          string
 	ServerPort         string
 	ServerIP           string
-	MaxMemory          uint32
+	MaxMemory          uint64
 	MaxMemoryPolicy    string
 	ConfigFiles        []string
 }
@@ -52,7 +52,7 @@ func (r *Redis) AddFlags(fs *pflag.FlagSet) {
 	fs.StringVar(&r.ConfigFileName, "c", RedisConfigFileDefault, "redis config file path")
 	fs.StringVar(&r.renameCommandsPath, "rename-command-path", RedisRenameCommandsDefaultPath, "Path to the folder where rename-commands option for redis are available")
 	fs.StringVar(&r.renameCommandsFile, "rename-command-file", RedisRenameCommandsDefaultFile, "Name of the file where rename-commands option for redis are available, disabled if empty")
-	fs.Uint32Var(&r.MaxMemory, "max-memory", RedisMaxMemoryDefault, "redis max memory")
+	fs.Uint64Var(&r.MaxMemory, "max-memory", RedisMaxMemoryDefault, "redis max memory")
 	fs.StringVar(&r.MaxMemoryPolicy, "max-memory-policy", RedisMaxMemoryPolicyDefault, "redis max memory eviction policy")
 	fs.StringVar(&r.ServerBin, "bin", RedisServerBinDefault, "redis server binary file name")
 	fs.StringVar(&r.ServerPort, "port", RedisServerPortDefault, "redis server listen port")
